Move generator package under cmd/genji
