+++ cmd/genji/generator/record.go
