+++ cmd/genji/generator/testdata/structs.generated.golden.go
