+++ cmd/genji/generator/testdata/structs.go
